factor out library
